+++ lib/image_optim/config.rb
@@ -16,21 +16,9 @@ class ImageOptim
     LOCAL_PATH = './.image_optim.yml'
 
     class << self
-      # Read config at GLOBAL_PATH if it exists, warn if anything is
-      # wrong
-      def global
-        read(GLOBAL_PATH)
-      end
-
-      # Read config at LOCAL_PATH if it exists, warn if anything is
-      # wrong
-      def local
-        read(LOCAL_PATH)
-      end
-
-    private
-
-      def read(path)
+      # Read options at path: expand path (warn on failure), return {} if file
+      # does not exist, read yaml, check if it is a Hash, deep symbolise keys
+      def read_options(path)
         begin
           full_path = File.expand_path(path)
         rescue ArgumentError => e
@@ -50,11 +38,12 @@ class ImageOptim
     end
 
     def initialize(options)
-      @options = [
-        Config.global,
-        Config.local,
-        HashHelpers.deep_symbolise_keys(options),
-      ].reduce do |memo, hash|
+      config_paths = options.delete(:config_paths) || [GLOBAL_PATH, LOCAL_PATH]
+
+      to_merge = config_paths.map{ |path| self.class.read_options(path) }
+      to_merge << HashHelpers.deep_symbolise_keys(options)
+
+      @options = to_merge.reduce do |memo, hash|
         HashHelpers.deep_merge(memo, hash)
       end
       @used = Set.new