+++ spec/image_optim/config_spec.rb
@@ -5,12 +5,11 @@ require 'image_optim/config'
 describe ImageOptim::Config do
   Config = ImageOptim::Config
 
-  before do
-    allow(Config).to receive(:global).and_return({})
-    allow(Config).to receive(:local).and_return({})
-  end
-
   describe 'assert_no_unused_options!' do
+    before do
+      allow(Config).to receive(:read_options).and_return({})
+    end
+
     it 'should not raise when no unused options' do
       config = Config.new({})
       config.assert_no_unused_options!
@@ -25,6 +24,10 @@ describe ImageOptim::Config do
   end
 
   describe 'nice' do
+    before do
+      allow(Config).to receive(:read_options).and_return({})
+    end
+
     it 'should be 10 by default' do
       config = Config.new({})
       expect(config.nice).to eq(10)
@@ -42,6 +45,10 @@ describe ImageOptim::Config do
   end
 
   describe 'threads' do
+    before do
+      allow(Config).to receive(:read_options).and_return({})
+    end
+
     it 'should be processor_count by default' do
       config = Config.new({})
       allow(config).to receive(:processor_count).and_return(13)
@@ -60,6 +67,10 @@ describe ImageOptim::Config do
   end
 
   describe 'for_worker' do
+    before do
+      allow(Config).to receive(:read_options).and_return({})
+    end
+
     Abc = Class.new do
       def self.bin_sym
         :abc
@@ -93,31 +104,46 @@ describe ImageOptim::Config do
     end
   end
 
+  describe 'config' do
+    it 'should read options from default locations' do
+      expect(Config).to receive(:read_options).
+        with(Config::GLOBAL_PATH).and_return(:a => 1, :b => 2, :c => 3)
+      expect(Config).to receive(:read_options).
+        with(Config::LOCAL_PATH).and_return(:a => 10, :b => 20)
+
+      config = Config.new(:a => 100)
+      expect(config.get!(:a)).to eq(100)
+      expect(config.get!(:b)).to eq(20)
+      expect(config.get!(:c)).to eq(3)
+      config.assert_no_unused_options!
+    end
+
+    it 'should not read options with empty config_paths' do
+      expect(Config).not_to receive(:read_options)
+
+      config = Config.new(:config_paths => [])
+      config.assert_no_unused_options!
+    end
+
+    it 'should read options from specified paths' do
+      expect(Config).to receive(:read_options).
+        with('/etc/image_optim.yml').and_return(:a => 1, :b => 2, :c => 3)
+      expect(Config).to receive(:read_options).
+        with('config/image_optim.yml').and_return(:a => 10, :b => 20)
+
+      config = Config.new(:a => 100, :config_paths => %w[
+        /etc/image_optim.yml
+        config/image_optim.yml
+      ])
+      expect(config.get!(:a)).to eq(100)
+      expect(config.get!(:b)).to eq(20)
+      expect(config.get!(:c)).to eq(3)
+      config.assert_no_unused_options!
+    end
+  end
+
   describe 'class methods' do
-    before do
-      allow(Config).to receive(:global).and_call_original
-      allow(Config).to receive(:local).and_call_original
-    end
-
-    describe 'global' do
-      it 'should call read with GLOBAL_PATH' do
-        expect(Config).to receive(:read).
-          with(Config::GLOBAL_PATH).and_return(:config => true)
-
-        expect(Config.global).to eq(:config => true)
-      end
-    end
-
-    describe 'local' do
-      it 'should call read with LOCAL_PATH' do
-        expect(Config).to receive(:read).
-          with(Config::LOCAL_PATH).and_return(:config => true)
-
-        expect(Config.local).to eq(:config => true)
-      end
-    end
-
-    describe 'read' do
+    describe 'read_options' do
       let(:path){ double(:path) }
       let(:full_path){ double(:full_path) }
 
@@ -127,7 +153,7 @@ describe ImageOptim::Config do
           with(path).and_raise(ArgumentError)
         expect(File).not_to receive(:file?)
 
-        expect(Config.send(:read, path)).to eq({})
+        expect(Config.read_options(path)).to eq({})
       end
 
       it 'should return empty hash if path is not a file' do
@@ -137,7 +163,7 @@ describe ImageOptim::Config do
         expect(File).to receive(:file?).
           with(full_path).and_return(false)
 
-        expect(Config.send(:read, path)).to eq({})
+        expect(Config.read_options(path)).to eq({})
       end
 
       it 'should return hash with deep symbolised keys from reader' do
@@ -152,7 +178,7 @@ describe ImageOptim::Config do
         expect(YAML).to receive(:load_file).
           with(full_path).and_return(stringified)
 
-        expect(Config.send(:read, path)).to eq(symbolized)
+        expect(Config.read_options(path)).to eq(symbolized)
       end
 
       it 'should warn and return an empty hash if reader returns non hash' do
@@ -164,7 +190,7 @@ describe ImageOptim::Config do
         expect(YAML).to receive(:load_file).
           with(full_path).and_return([:config])
 
-        expect(Config.send(:read, path)).to eq({})
+        expect(Config.read_options(path)).to eq({})
       end
 
       it 'should warn and return an empty hash if reader raises exception' do
@@ -176,7 +202,7 @@ describe ImageOptim::Config do
         expect(YAML).to receive(:load_file).
           with(full_path).and_raise
 
-        expect(Config.send(:read, path)).to eq({})
+        expect(Config.read_options(path)).to eq({})
       end
     end
   end